chore: disable ssr for browser-only apps
